% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_extent.R
\name{load_extent_bbox}
\alias{load_extent_bbox}
\alias{load_extent_vectfile}
\alias{load_extent_vectfile_pa}
\alias{load_extent_vectfile_mask}
\alias{load_extent_draw}
\alias{load_extent_forest}
\title{Insert an extent}
\usage{
load_extent_bbox()

load_extent_vectfile()

load_extent_vectfile_pa()

load_extent_vectfile_mask()

load_extent_draw(extent_ns_name)

load_extent_forest()
}
\arguments{
\item{extent_ns_name}{Name of the namespace to be used}
}
\description{
Internal functions and modal dialogs to specify an extent
in the GUI.
}
\note{
License: GPL 3.0
}
\author{
Pascal Obstetar (2018) \email{pascal.obstetar@gmail.com}
}
