% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Prediction with preprocess Sentinel-2 images calculated}
\usage{
prediction(param_list = NULL)
}
\arguments{
\item{param_list}{(optional) List of input parameters:
it can be both an R list or the path of a JSON file.
If some parameters are passed both as elements of \code{param_list}
and as function arguments, the values passed as function
arguments are considered.
If some parameters are missing in \code{param_list} and are not
provided as arguments, default values will be used.
Use the function \code{\link[=s2_gui]{s2_gui()}} to create a complete list of
parameters.
If \code{param_list} is NULL (default), values given with the
parameters below (or default values for parameters not
provided) are used.}

\item{path_indices}{(optional) Path of the directory in which files of
spectral indices are searched and/or generated.
If not provided (default), \code{path_out} is used.}

\item{sigma}{(optional) Default value 0.999, this is the value for extract
point from raster file predict.}

\item{parallel}{(optional) Logical or integer: if TRUE (default), some
functions (\link{sen2cor}, \link{s2_mask} and \link{s2_calcindices} for now)
are executed using multiple cores in order to speed up the execution.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE, the processing chain is forced to run with a single core
(this can be useful if multiple \link{cnes} instances are run in parallel).
This argument can be set only in commandline mode, not using the GUI.}

\item{use_python}{(optional) Logical: if TRUE (default), the presence of
python in the system is checked before running the function;
if FALSE, this is skipped. Setting this to FALSE can bge useful on
systems with problems with python, when \code{\link[=theia2r]{theia2r()}} is intended
to be used only for processing existing SAFE files (python is required
in any case to download SAFE).}

\item{tmpdir}{(optional) Path where intermediate files will be created.
Default is a temporary directory (unless \code{outformat = "VRT"}: in this case,
default is a subdirectory named ".vrt" within \code{path_out}).}
}
\value{
A vector with the paths of the files which were created (excluded
the temporary files); NULL otherwise.
The vector includes two attributes:
\itemize{
\item \code{cloudcovered} with the list of imags not created due to the higher
percentage of cloud covered pixels;
\item \code{missing} with the list of imags not created due to other reasons.
}
}
\description{
The function is a wrapper to perform the entire
processing chain to find, download and pre-process Sentinel-2
data and produce prediction of Presence/Absence of a phenome.
Input is a set of parameters that can be passed with a
list or file (parameter \code{param_list}) or singularly (see the
descriptions of all the other parameters).
}
