% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Find, download and preprocess Sentinel-2 images}
\usage{
preprocessing(param_list = NULL)
}
\arguments{
\item{param_list}{(optional) List of input parameters:
it can be both an R list or the path of a JSON file.
If some parameters are passed both as elements of \code{param_list}
and as function arguments, the values passed as function
arguments are considered.
If some parameters are missing in \code{param_list} and are not
provided as arguments, default values will be used.
Use the function \code{\link[=s2_gui]{s2_gui()}} to create a complete list of
parameters.
If \code{param_list} is NULL (default), values given with the
parameters below (or default values for parameters not
provided) are used.}

\item{online}{(optional) Logical: TRUE (default) to search for available
products on SciHub (and download if needed); FALSE to work
only with already downloaded SAFE products.}

\item{api}{Path of the text file containing credentials
of cnes account. If NA (default) the default credentials
(username "email", password "password") will be used.}

\item{downloader}{(optional) Character value corresponding to the executable
which should be used to download SAFE products. It could be one among
"wget" (default) and "aria2".
If aria2 is not installed, Wget will be used instead.}

\item{overwrite_product}{(optional) Logical: TRUE to overwrite existing
products with products found online or manually corrected,
FALSE (default) to skip download for products already existing.}

\item{timewindow}{(optional) Temporal window for querying: Date object
of length 1 (single day) or 2 (time window). Default is NA, meaning that
no filters are used if online = FALSE, and all found images are processed;
if online = TRUE, last 90 days are processed.
Is it possible to pass also integer (or difftime) values, which are
interpreted as the last n days.}

\item{timeperiod}{(optional) Character:
\itemize{
\item "full" (default) means that all
the images included in the time window are considered;
\item "seasonal" means that only the single seasonal periods in the
window are used (i.e., with a time window from 2015-06-01 to
2017-08-31, the periods 2015-06-01 to 2015-08-31, 2016-06-01
to 2016-08-31 and 2017-06-01 to 2017-08-31 are considered).
}}

\item{extent}{(optional) Spatial extent on which to clip products (it can
be both the path of a vector file or a geoJSON).
Default is NA for offline mode (meaning no extent:
all found tiles are entirely used); in online mode, a sample extent is used
as default.}

\item{list_rgb}{(optional) Character vector with the values of the
RGB images to be produced.
Images are in the form xRGBrgb, when:
\itemize{
\item x is B (if source is BOA) or T (is source is TOA);
\item r g and b are the the number of the bands to be used respectively
for red, green and blue, in hexadecimal format.
Notice that this is the \href{https://earth.esa.int/web/sentinel/user-guides/sentinel-2-msi/resolutions/spatial}{actual number name of the bands}:
so, to use i.e. BOA band 11 (1610nm) use the value "b", even if band 11 is
the 10th band of a BOA product (because band 10 is missing).
Default is no one (NA).
}}

\item{index_source}{(optional) Character value: if "BOA" (default), indices
are computed from BOA values; if "TOA", non corrected reflectances
are instead used (be careful to use this setting!).}

\item{mask_type}{(optional) Character value which determines the categories
in the Surface Classification Map to be masked (see \code{\link[=s2_mask]{s2_mask()}}
for the accepted values). Default (NA) is not to mask.}

\item{max_mask}{(optional) Numeric value (range 0 to 100), which represents
the maximum percentage of allowed masked surface (by clouds or any other
type of mask chosen with argument \code{mask_type}) for producing outputs.
Images with a percentage of masked surface greater than \code{max_mask}%
are not processed (the list of expected output files which have not been
generated is returned as an attribute, named "skipped").
Default value is 80.
Notice that the percentage is computed on non-NA values (if input images
had previously been clipped and masked using a polygon, the percentage is
computed on the surface included in the masking polygons).}

\item{mask_smooth}{(optional) Numeric positive value: the smoothing radius
(expressed in unit of measure of the output projection, typically metres)
to be applied to the cloud mask by function \link{s2_mask}.}

\item{mask_buffer}{(optional) Numeric value: the buffering radius
(expressed in unit of measure of the output projection, typically metres)
to be applied to the cloud mask by function \link{s2_mask}.
Default value (0) means that no buffer is applied; a positive value causes
an enlargement of the masked area; a negative value cause a reducement.}

\item{resampling}{(optional) Resampling method (one of the values supported
by \code{gdal_translate}: "near" (default), "bilinear", "cubic",
"cubicspline", "lanczos", "average" or "mode").}

\item{resampling_scl}{(optional) Resampling method for categorical products
(for now, only SCL): one among "near" (default) and "mode".}

\item{outformat}{(optional) Format of the output file (in a
format recognised by GDAL). Default is "GTiff".}

\item{rgb_outformat}{(optional) Format of the output RGB products (in a
format recognised by GDAL). Default is "GTiff".}

\item{index_datatype}{(optional) Numeric datatype of the ouptut
spectral indices.}

\item{compression}{(optional) In the case GTiff is chosen as
output format, the compression indicated with this parameter is
used (default is "DEFLATE").}

\item{rgb_compression}{(optional) In the case GTiff is chosen as
output format for RGB products, the compression indicated
with this parameter is used (default is "DEFLATE").
In the cases GTiff or JPEG are chosen as output format for RGB products,
this parameter can also be a 1-100 integer value, which is interpreted
as the compression level for a JPEG compression.}

\item{overwrite}{(optional) Logical value: should existing output
files be overwritten? (default: FALSE).}

\item{path_theia_landsat}{(optional) Path of the directory in which THEIA LANDSAT
products are searched and/or downloaded. If not provided (default), a
temporary directory is used.}

\item{path_theia_spotworldheritage}{(optional) Path of the directory in which THEIA SpotWorldHeritage
products are searched, downloaded and/or generated. If not provided
(default), a temporary directory is used.}

\item{path_theia_sentinel2}{(optional) Path of the directory in which THEIA Sentinel2
products are searched and/or downloaded. If not provided (default), a
temporary directory is used.}

\item{path_theia_snow}{(optional) Path of the directory in which THEIA Snow
products are searched, downloaded and/or generated. If not provided
(default), a temporary directory is used.}

\item{path_theia_venus}{(optional) Path of the directory in which THEIA VENUS
products are searched, downloaded and/or generated. If not provided
(default), a temporary directory is used.}

\item{path_out}{(optional) Path of the directory in which Sentinel-2
output products are searched and/or generated.
If not provided (default), a temporary directory is used.}

\item{path_rgb}{(optional) Path of the directory in RGB products
are searched and/or generated.
If not provided (default), \code{path_out} is used.}

\item{path_indices}{(optional) Path of the directory in which files of
spectral indices are searched and/or generated.
If not provided (default), \code{path_out} is used.}

\item{thumbnails}{(optional) Logical: if TRUE (default), a thumbnail is
added for each product created. Thumbnails are JPEG or PNG georeferenced
small images (width or height of 1024 pixels) with default colour palettes
(for more details, see the help window in the GUI). They are placed in
a subdirectory of the products names "thumbnails".
If FALSE, they are not created.}

\item{parallel}{(optional) Logical or integer: if TRUE (default), some
functions (\link{sen2cor}, \link{s2_mask} and \link{s2_calcindices} for now)
are executed using multiple cores in order to speed up the execution.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE, the processing chain is forced to run with a single core
(this can be useful if multiple \link{cnes} instances are run in parallel).
This argument can be set only in commandline mode, not using the GUI.}

\item{use_python}{(optional) Logical: if TRUE (default), the presence of
python in the system is checked before running the function;
if FALSE, this is skipped. Setting this to FALSE can bge useful on
systems with problems with python, when \code{\link[=theia2r]{theia2r()}} is intended
to be used only for processing existing SAFE files (python is required
in any case to download SAFE).}

\item{tmpdir}{(optional) Path where intermediate files will be created.
Default is a temporary directory (unless \code{outformat = "VRT"}: in this case,
default is a subdirectory named ".vrt" within \code{path_out}).}

\item{rmtmp}{(optional) Logical: should temporary files be removed?
(Default: TRUE). \code{rmtmp} is forced to \code{FALSE} if \code{outformat = "VRT"}.}

\item{log}{(optional) Character string with the path where the package
messages will be redirected.
Default (NA) is not to redirect (use standard output).
A two-length character with tho paths (which can also coincide)
can be used to redirect also the output: in this case, the first path}

\item{product}{Product}

\item{theiacollection}{Theia collection}

\item{pepscollection}{Peps collection}

\item{theiaplatformlandsat}{Platform THEIA landsat}

\item{theiaplatformspotworldheritage}{Platform THEIA Spot World Heritage}

\item{theiaplatformsentinel}{Platform THEIA SENTINEL 2}

\item{theiaplatformvenus}{Platform THEIA VENUS}

\item{extent_name}{Extent name}

\item{list_indices_checked}{List indices checked}

\item{reference_path}{is the path for messages, the second one for the output.}
}
\value{
A vector with the paths of the files which were created (excluded
the temporary files); NULL otherwise.
The vector includes two attributes:
\itemize{
\item \code{cloudcovered} with the list of imags not created due to the higher
percentage of cloud covered pixels;
\item \code{missing} with the list of imags not created due to other reasons.
}
}
\description{
The function is a wrapper to perform the entire
processing chain to find, download and pre-process Sentinel-2
data. Input is a set of parameters that can be passed with a
list or file (parameter \code{param_list}) or singularly (see the
descriptions of all the other parameters).
}
