% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_download.R
\name{product_download}
\alias{product_download}
\title{Download THEAI or PEPS products.}
\usage{
product_download(spatial_extent = NULL, tile = NULL, orbit = NULL,
  time_interval = NULL, time_period = "full", level = "L2A",
  platform = "SENTINEL2A", maxcloud = 101, collection = "SENTINEL",
  api = NA, writedir = ".")
}
\arguments{
\item{spatial_extent}{Spatiam extent}

\item{tile}{Tile}

\item{orbit}{Orbit}

\item{time_interval}{Time interval}

\item{time_period}{Time period}

\item{level}{Level}

\item{platform}{Platform}

\item{maxcloud}{Max cloud cover}

\item{collection}{Collection}

\item{api}{Path of the 'api.txt' file containing credentials
of cnes account. If NA (default) the default credentials}

\item{writedir}{Write dir
(username 'user', password 'user') will be used.}
}
\description{
The function downloads product.
(the content must be a URL, and the name the product name).
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
# Download the whole product
product_download(prodlist, outdir = tempdir())

# Download a serie of products
pos <- st_sfc(st_point(c(12.0, 44.8)), crs = st_crs(4326))
time_window <- as.Date(c("2017-05-01", "2017-07-30"))
example_theia_list <- s2_list(spatial_extent = pos, time_interval = time_window)
product_download(example_theia_list, outdir = tempdir())
}

}
\author{
Pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
