% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2rgb.R
\name{raster2rgb}
\alias{raster2rgb}
\title{Produce an RGB image from a singleband raster file.}
\usage{
raster2rgb(in_rast, out_file = NULL, palette = "generic_ndsi",
  tmpdir = NA, scaleRange = NA)
}
\arguments{
\item{in_rast}{Path of the input multiband raster.}

\item{out_file}{(optional) Path of the output RGB JPEG image; if NULL
(default), a RasterLayer will be returned.}

\item{palette}{Path of the palette file to be used (cpt or txt),
or character value of a builtin palette (the default
"generic_ndsi").}

\item{tmpdir}{(optional) Path where intermediate files (VRT) will be created.
Default is a temporary directory.}
}
\value{
The path of the output image; alternatively, the output image
as RasterLayer (if \code{out_rast = NULL}).
}
\description{
Internal function to create JPEG or PNG images from a
singleband raster file. This function is used by \link{s2_thumbnails},
and it will be exported when it would be more generalised.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}, Pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
