% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_rgb.R
\name{s2_rgb}
\alias{s2_rgb}
\title{Create RGB images from S2 reflectance products.}
\usage{
s2_rgb(infiles, rgblist, tilesdir, extent, mosaicdir, project,
  outdir = ".", subdirs = NA, tmpdir = NA, rmtmp = TRUE,
  format = NA, compress = "DEFLATE", vrt_rel_paths = NA,
  parallel = FALSE, resolution = 10, overwrite = FALSE)
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the theia2r naming convention
(\link{safe_shortname}).}

\item{rgblist}{rgblist}

\item{tilesdir}{Directory of tiles}

\item{mosaicdir}{Directory of mosaic}

\item{outdir}{(optional) Full name of the existing output directory
where the files should be created. Default is the same directory of
input reflectance files. # FIXME use a subdir with product name}

\item{subdirs}{(optional) Logical: if TRUE, different indices are
placed in separated \code{outfile} subdirectories; if FALSE, they are placed in
\code{outfile} directory; if NA (default), subdirectories are created only if
more than a single spectral index is required.}

\item{tmpdir}{(optional) Path where intermediate files (VRT) will be created.
Default is a temporary directory.}

\item{rmtmp}{(optional) Logical: should temporary files be removed?
(Default: TRUE)}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is the same format of input images
(or 'GTiff' in case of VRT input images).}

\item{compress}{(optional) In the case a GTiff format is
present, the compression indicated with this parameter is used.}

\item{vrt_rel_paths}{Directory vrt}

\item{parallel}{(optional) Logical: if TRUE, the function is run using parallel
processing, to speed-up the computation for large rasters.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE (default), single core processing is used.}

\item{resolution}{Resolution}

\item{overwrite}{(optional) Logical value: should existing thumbnails be}

\item{overwritten?}{(default: TRUE)}
}
\value{
A vector with the names of the created images.
}
\description{
Function to create RGB images from Sentinel-2 reflectances.
natural: B4, B3, B2
near infrarouge: B8, B4, B3
far infrarouge: B12, B8, B4
vegetation: B11, B8, B2
}
\note{
License: GPL 3.0
}
\author{
pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
