% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_translate.R
\name{s2_translate}
\alias{s2_translate}
\title{Make a translate of each band of S2 tiles with the same date and orbit}
\usage{
s2_translate(infiles, mosaicdir, tilesdir, outdir = ".", subdirs = NA,
  tmpdir = NA, rmtmp = TRUE, format = NA, compress = "DEFLATE",
  vrt_rel_paths = NA, resolution = 10, overwrite = FALSE)
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted to a format managed by
GDAL.}

\item{mosaicdir}{Directory of mosaic}

\item{tilesdir}{Directory of tiles}

\item{outdir}{(optional) Full name of the output directory where
the files should be created (default: current directory).
\code{outdir} can bot be an existing or non-existing directory (in the
second case, its parent directory must exists).
If it is a relative path, it is expanded from the common parent
directory of \code{infiles}.}

\item{subdirs}{Subdirectory}

\item{tmpdir}{(optional) Path where intermediate files (VRT) will be created.
Default is a temporary directory.}

\item{rmtmp}{(optional) Logical: should temporary files be removed?
(Default: TRUE).
This parameter takes effect only if the output files are not VRT
(in this case temporary files cannot be deleted, because rasters of source
bands are included within them).}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is to maintain each input format.}

\item{compress}{(optional) In the case a GTiff format is
present, the compression indicated with this parameter is used.}

\item{vrt_rel_paths}{(optional) Logical: if TRUE (default on Linux),
the paths present in the VRT output file are relative to the VRT position;
if FALSE (default on Windows), they are absolute.
This takes effect only with \code{format = "VRT"}.}

\item{resolution}{Resolution
overwritten? (default: FALSE)}

\item{overwrite}{Logical value: should existing output files be}
}
\value{
A vector with the names of the merged products (just created or
already existing).
}
\description{
The function mosaic the input Sentinel-2 files with
the same date, orbit number, product type and file format.
Outputs are a set of products in the same format of corresponding
input files.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}, pascal Obstetar, (2019) \email{pascal.obstetar@gmail.com}
}
